Impl ValueResolver for constant
