+++ annotation/src/main/java/online/sharedtype/SharedType.java
@@ -259,47 +259,23 @@ public @interface SharedType {
 
     /**
      * Mark enum value. By default, enum value is the enum constant name. The enum value must be literals (e.g. 1, "a", true) in enum constant expressions.
-     * <p>
-     * When placed on:
-     *  <ul>
-     *      <li>Constructor parameter - the literal value served to this parameter from enum constant expressions will be used.</li>
-     *      <li>Field - the constructor parameter with the same name and type will be used as if constructor parameter is annotated.</li>
-     *  </ul>
-     * <p>
-     * Below are some valid examples:
-     * </p>
+     * <b>Note: When there are multiple enum constant constructor parameters, the value is resolved by field order.</b>
+     * If the constructor parameter order is different from the field order, value will not be resolved correctly.
      * <pre>
+     * Example:
      * {@code
      * enum Enum {
      *   A(1), B(2);
      *
      *   @SharedType.EnumValue
      *   private final int value;
-     *
-     *   Enum(int value) {
-     *     this.value = value;
-     *   }
-     * }
-     * }
-     * </pre>
-     * is equivalent to:
-     * <pre>
-     * {@code
-     * enum Enum {
-     *   A(1), B(2);
-     *
-     *   private final int value;
-     *
-     *   Enum(@SharedType.EnumValue int value) {
-     *     this.value = value;
-     *   }
      * }
      * }
      * </pre>
      * <br>
      * Additional annotation types can be configured via global properties.
      */
-    @Target({ElementType.FIELD, ElementType.PARAMETER})
+    @Target({ElementType.FIELD})
     @Retention(RetentionPolicy.CLASS)
     @interface EnumValue {
     }
