+++ processor/src/main/java/online/sharedtype/processor/parser/EnumTypeDefParser.java
@@ -1,33 +1,26 @@
 package online.sharedtype.processor.parser;
 
-import com.sun.source.tree.ExpressionTree;
-import com.sun.source.tree.LiteralTree;
-import com.sun.source.tree.NewClassTree;
 import com.sun.source.tree.Tree;
-import com.sun.source.tree.VariableTree;
 import lombok.RequiredArgsConstructor;
+import online.sharedtype.processor.context.Config;
+import online.sharedtype.processor.context.Context;
 import online.sharedtype.processor.domain.ConcreteTypeInfo;
 import online.sharedtype.processor.domain.DependingKind;
 import online.sharedtype.processor.domain.EnumDef;
 import online.sharedtype.processor.domain.EnumValueInfo;
 import online.sharedtype.processor.domain.TypeDef;
 import online.sharedtype.processor.domain.TypeInfo;
-import online.sharedtype.processor.context.Config;
-import online.sharedtype.processor.context.Context;
 import online.sharedtype.processor.parser.type.TypeContext;
 import online.sharedtype.processor.parser.type.TypeInfoParser;
-import online.sharedtype.processor.support.exception.SharedTypeInternalError;
+import online.sharedtype.processor.parser.value.ValueResolver;
 
 import javax.lang.model.element.Element;
 import javax.lang.model.element.ElementKind;
-import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.TypeElement;
 import javax.lang.model.element.VariableElement;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-
-import static online.sharedtype.processor.domain.Constants.STRING_TYPE_INFO;
 
 /**
  * Literal values are parsed via {@link Tree} API. It has limitations, see the documentation for more details.
@@ -38,6 +31,7 @@ import static online.sharedtype.processor.domain.Constants.STRING_TYPE_INFO;
 final class EnumTypeDefParser implements TypeDefParser {
     private final Context ctx;
     private final TypeInfoParser typeInfoParser;
+    private final ValueResolver valueResolver;
 
     @Override
     public List<TypeDef> parse(TypeElement typeElement) {
@@ -53,14 +47,9 @@ final class EnumTypeDefParser implements TypeDefParser {
         List<? extends Element> enclosedElements = typeElement.getEnclosedElements();
         List<VariableElement> enumConstantElems = new ArrayList<>(enclosedElements.size());
 
-        EnumValueMarker enumValueMarker = new EnumValueMarker(ctx, config);
         for (Element enclosedElement : enclosedElements) {
             if (enclosedElement.getKind() == ElementKind.ENUM_CONSTANT) {
                 enumConstantElems.add((VariableElement) enclosedElement);
-            } else if (enclosedElement.getKind() == ElementKind.CONSTRUCTOR) {
-                enumValueMarker.parseConstructor((ExecutableElement) enclosedElement);
-            } else if (ctx.isAnnotatedAsEnumValue(enclosedElement)) {
-                enumValueMarker.setField((VariableElement) enclosedElement);
             }
         }
 
@@ -69,141 +58,25 @@ final class EnumTypeDefParser implements TypeDefParser {
             .simpleName(config.getSimpleName())
             .annotated(config.isAnnotated())
             .build();
-        enumDef.components().addAll(
-            enumValueMarker.marked() ? parseEnumConstants(typeElement, enumConstantElems, enumValueMarker, enumDef) : useEnumConstantNames(enumConstantElems)
-        );
+        enumDef.components().addAll(parseEnumConstants(typeElement, enumConstantElems));
         TypeInfo typeInfo = typeInfoParser.parse(typeElement.asType(), TypeContext.builder().typeDef(enumDef).dependingKind(DependingKind.SELF).build());
         enumDef.linkTypeInfo((ConcreteTypeInfo) typeInfo);
         ctx.getTypeStore().saveConfig(enumDef.qualifiedName(), config);
         return Collections.singletonList(enumDef);
     }
 
-    private static List<EnumValueInfo> useEnumConstantNames(List<VariableElement> enumConstants) {
+    private List<EnumValueInfo> parseEnumConstants(TypeElement enumTypeElement, List<VariableElement> enumConstants) {
         List<EnumValueInfo> res = new ArrayList<>(enumConstants.size());
         for (VariableElement enumConstant : enumConstants) {
             String name = enumConstant.getSimpleName().toString();
-            res.add(new EnumValueInfo(name, STRING_TYPE_INFO, name));
-        }
-        return res;
-    }
-
-    private List<EnumValueInfo> parseEnumConstants(TypeElement enumTypeElement,
-                                                   List<VariableElement> enumConstants,
-                                                   EnumValueMarker enumValueMarker,
-                                                   EnumDef enumDef) {
-        List<EnumValueInfo> res = new ArrayList<>(enumConstants.size());
-        TypeInfo valueTypeInfo = typeInfoParser.parse(
-            enumValueMarker.enumValueVariableElem.asType(),
-            TypeContext.builder().typeDef(enumDef).dependingKind(DependingKind.ENUM_VALUE).build()
-        );
-        int ctorArgIdx = enumValueMarker.matchAndGetConstructorArgIdx();
-        if (ctorArgIdx < 0) {
-            return Collections.emptyList();
-        }
-        for (VariableElement enumConstant : enumConstants) {
-            String name = enumConstant.getSimpleName().toString();
-            Tree tree = ctx.getTrees().getTree(enumConstant);
-            if (tree instanceof VariableTree) {
-                VariableTree variableTree = (VariableTree) tree;
-                Object value = resolveValue(enumTypeElement, variableTree, ctorArgIdx);
-                if (value != null) {
-                    res.add(new EnumValueInfo(name, valueTypeInfo, value));
-                }
-            } else if (tree == null) {
-                ctx.error(enumConstant, "Literal value cannot be parsed from enum constant: %s of enum %s, because source tree from the element is null." +
-                    " This could mean at the time of the annotation processing, the source tree was not available." +
-                    " Is this class from a dependency jar/compiled class file? Please refer to the documentation for more information.",
-                    enumConstant, enumTypeElement);
+            Object value = valueResolver.resolve(enumConstant, enumTypeElement);
+            if (value != null) {
+                TypeInfo valueTypeInfo = ctx.getTypeStore().getTypeInfo(value.getClass().getCanonicalName(), Collections.emptyList());
+                res.add(new EnumValueInfo(name, valueTypeInfo, value));
             } else {
-                throw new SharedTypeInternalError(String.format(
-                    "Unsupported tree during parsing enum %s, kind: %s, tree: %s, element: %s", enumTypeElement, tree.getKind(), tree, enumConstant));
+                ctx.warn(enumConstant, "Cannot resolve value for enum constant %s", name);
             }
         }
         return res;
-    }
-
-    private Object resolveValue(TypeElement enumTypeElement, VariableTree tree, int ctorArgIdx) {
-        ExpressionTree init = tree.getInitializer();
-        if (init instanceof NewClassTree) {
-            NewClassTree newClassTree = (NewClassTree) init;
-            try {
-                ExpressionTree argTree = newClassTree.getArguments().get(ctorArgIdx);
-                if (argTree instanceof LiteralTree) {
-                    LiteralTree argLiteralTree = (LiteralTree) argTree;
-                    return argLiteralTree.getValue();
-                } else {
-                    ctx.error(enumTypeElement, "Unsupported argument in enum type %s: %s in %s, argIndex: %s. Only literals are supported as enum value.",
-                        enumTypeElement, argTree, tree, ctorArgIdx);
-                    return null;
-                }
-            } catch (IndexOutOfBoundsException e) {
-                throw new SharedTypeInternalError(String.format(
-                    "Initializer in enum %s has incorrect number of arguments: %s in tree: %s, argIndex: %s", enumTypeElement, init, tree, ctorArgIdx));
-            }
-        }
-        throw new SharedTypeInternalError(String.format("Unsupported initializer in enum %s: %s in tree: %s", enumTypeElement, init, tree));
-    }
-
-    @RequiredArgsConstructor
-    static final class EnumValueMarker {
-        private final Context ctx;
-        private final Config config;
-        private List<String> constructorArgNames = Collections.emptyList();
-        private int constructorArgIdx = -1;
-        private VariableElement enumValueVariableElem;
-
-        void parseConstructor(ExecutableElement constructor) {
-            List<? extends VariableElement> parameters = constructor.getParameters();
-            constructorArgNames = new ArrayList<>(parameters.size());
-            for (int i = 0, n = parameters.size(); i < n; i++) {
-                VariableElement arg = parameters.get(i);
-                constructorArgNames.add(arg.getSimpleName().toString());
-                if (ctx.isAnnotatedAsEnumValue(arg)) {
-                    setField(arg);
-                    this.constructorArgIdx = i;
-                }
-            }
-        }
-
-        void setField(VariableElement enumValueVariableElem) {
-            if (this.enumValueVariableElem != null) {
-                ctx.error(enumValueVariableElem, "Enum %s has multiple fields annotated as enum value, only one field or constructor parameter is allowed, found on %s and %s",
-                    config.getQualifiedName(), this.enumValueVariableElem, enumValueVariableElem);
-            } else {
-                this.enumValueVariableElem = enumValueVariableElem;
-            }
-        }
-
-        /**
-         * Return -1 if no constructor parameter can be matched
-         */
-        int matchAndGetConstructorArgIdx() {
-            if (constructorArgIdx >= 0) {
-                return constructorArgIdx;
-            }
-
-            for (int i = 0, n = constructorArgNames.size(); i < n; i++) {
-                if (constructorArgNames.get(i).equals(enumValueVariableElem.getSimpleName().toString())) {
-                    return i;
-                }
-            }
-
-            String lombokSuggestion = "";
-            if (constructorArgNames.isEmpty()) {
-                lombokSuggestion = "The discovered constructor has 0 parameter, if Lombok is used to generate the constructor," +
-                    " please ensure annotation processing of Lombok is executed before SharedType. Or add explicit constructor." +
-                    " Later version of SharedType may infer constructor parameter position by field position without an explicit constructor.";
-            }
-
-            ctx.error(enumValueVariableElem, "Enum %s has a field annotated as enum value, but no constructor parameter can be matched."
-                    + lombokSuggestion
-                    + " May refer to the documentation on how to correctly mark enum value.",
-                config.getQualifiedName());
-            return -1;
-        }
-
-        boolean marked() {
-            return enumValueVariableElem != null;
-        }
     }
 }
